Add Verbose logging and improve logging in Invoke-FalconAuth and Get-ResourceContent functions
This commit adds a new parameter called `Verbose` to the `Write-FalconLog` function in the `falcon_windows_install.ps1` file. It also adds a new function called `Write-VerboseLog` that writes verbose logs to the console and log file. Additionally, the `Invoke-FalconAuth` and `Get-ResourceContent` functions were updated to log more information. These changes improve the logging capabilities of the script and make it easier to troubleshoot issues.