fix: allow sorting for kpagent (#232)
* fix: allow sorting for kpagent

This commit introduces a new function to manage listing tags and formatting them appropriately.
Prior to this, kpagent would return a list of tags that were in no particular order. Since kpagent
is a special child, we shall continue to treat it differently ;0. I've also added some refactoring
with the new function to assist --list-tags and making sure the latest sensor is downloaded.

* chore: update .gitignore

Ignore changes in .vscode directory