feat: add support for listing tags for podman/skopeo
This commit refactors the code to allow any container tool to list tags and take
advantage of the new list_tags function. This simplifies the case for listing tags
as well as downloading tags. Since a container tool is no longer needed to query
tags - we can omit those case statements.