feat: Add proxy to migrate (#248)
* add proxy params + use same log writing functions

The log writing functions should be the same in the migrate script as they are in the
install and uninstall scripts.

* ignore vagrantfile

* update logic to support new proxy changes

The install portion now following the install script logic. The new functions have been added.
The uninstall portion now supports proxy. TBD - Add proxy changes to actual uninstall script.

* update formatting