fix(pwsh_uninstall): fixes bug with removehost due to non-string being passed when exception caught (#269)
* fix(pwsh_uninstall): fixes bug with removehost due to non-string being passed when exception caught

Fixes #268

The issue here was anytime an exception occured when the API was called, the catch block
was sending an Exception Object to the Write-VerboseLog, which didn't know how to handle that
and would treat it as a JSON object, which would put it into an infinite loop. The fix was
to ensure that we were sending in the $_.Exception.Message which is a string.

* docs(pwsh): remove $true from RemoveHost as this does not work

The existing description of how to use the `RemoveHost` flag is not correct since we
are using a Switch, and not a Bool. So if you pass in $true, you might accidentally pass
that in as a arg to MaintenanceToken

* chore: add new ci workflow for testing uninstall script pwsh

* chore: minor ci workflow updates for pwsh