fix(bash_install): fix issue when installing with ubuntu 14
Fixes #273

This issue presents itself with very old versions of Apt < 1.1. To address this,
the traditional way to install a local deb package was to use dpkg first, which
would then error if any deps are needed, which is true in our case. Then afterwards
you could install all with apt-get install -f flag.