cmm: minor logic updates and doc updates
This takes some of the logic implemented in the uninstall script to deal
with ensuring that no api calls are made if a user passes in an option
like -h to the script. By wrapping those calls in funcs and only calling
them when appropriate, the script will run accordingly.