fix(container-pull): fix issue with artifactory incorrect tags + add version info
This PR fixes a current issue affecting Artifactory registries and
multi-arch images in which the digest of the images are being displayed
as tags. However, this poses a problem become the digest is not a valid
tag format. A recent CVE exposed this and now tools such as Skopeo and
Podman have patched this which is causing the listing tag issue with
Artifactory.

Also added a better way to get the version of the script currently being
used.