feat(container-pull): auto-detect container tool when not specified (#317)
Fixes #316

This PR adds the ability to auto-detect the container runtime tool when
not explicitly specified. Prior to this, it would simply default to
docker without checking to see if any other tools were installed. Now we
go in order -> Docker -> Podman -> Skopeo in the event that more than
one tool is installed.