fix(container-pull): update scopes and better error handling (#323)
* fix(container-pull): update scopes and better error handling

This PR cleans up existing API scopes and makes it easier to know what
scopes are needed depending on the sensor type. This allows for better
support of least privilege. Also added better handling of catching
errors associated with 403 API scope issues.

* docs: minor updates to the prereqs section

* chore: found minimum scopes for just pulling

This better emphasizes the least amount of scopes needed to simply
perform what the script is meant to do: pull down the respective images.
Anything outside of that would be beyond the scope of this script and
should be handled by another API credential.

* fix: admonition was incorrect - fixed formatting to make it work