fix(container-pull): fixes issue with passing full sensor version (#347)
The sensor version gets translated from UPPER:lower before being used. I
believe this was to do further processing on it and to ensure some form
of consistent format. Either way, the grep of the SENSOR_VERSION
variable needs to be case insensitive - allows a user to pass in a
sensor version (whether it be x.y.z-patch or the full name).

Fixes #340